+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/package.json
@@ -14,6 +14,7 @@
     "react-dom": "^16.8.6",
     "react-dropdown": "^1.6.4",
     "react-dropdown-select": "^3.0.0",
+    "react-responsive-carousel": "^3.1.49",
     "react-router-dom": "^5.0.0",
     "react-scripts": "3.0.1",
     "react-select": "^2.4.3",
@@ -39,5 +40,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "react-image-resizer": "^1.3.0"
   }
 }
