+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/Components/PopularEvents.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import '../App.scss';
+
+class PopularEvents extends React.Component {
+
+    render () {
+        return (
+            <section>
+                <article className="App-row-elements">
+                    <h2>Eventos Populares:</h2>
+                </article>
+                <article className="App-column-elements">
+                    
+                </article>
+            </section>
+        );
+    }
+  }
+
+export default PopularEvents;
