+++ bogota-abril-2019/final_project/Álvaro David Pongutá/nomadsnearu/src/Components/ViewEvent.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import '../App.scss';
 import firebase from '../Firebase';
+import Button from '../Atoms/Button';
 
 class ListActivities extends React.Component {
 
@@ -23,6 +24,12 @@ class ListActivities extends React.Component {
 
 class ViewEvent extends React.Component {
 
+    constructor(props) {
+        super(props);
+
+        this.assistToAnEvent = this.assistToAnEvent.bind(this);
+    }
+
     state = {
         actualEvent: ''
     };
@@ -33,6 +40,49 @@ class ViewEvent extends React.Component {
         }))
     }
 
+    assistToAnEvent() {
+
+        if(firebase.auth().currentUser != null){
+            let currentUID = firebase.auth().currentUser.uid;
+
+            let actualEventId = this.props.match.params.eventId;
+
+            //Mirar si en el evento hay cupos
+            firebase.database().ref('/events/' + actualEventId).once('value').then(event => {
+                if (event.val().numberOfRemainingAsisstants > 0){
+                    //Revisar que la persona no esté asistiendo al evento
+                    let assistingToEvent = false;
+                    firebase.database().ref('/eventsXusers').once('value', (snapshot) => {
+                        snapshot.forEach(snap => {
+                            if(snap.child(currentUID).key === currentUID){
+                                if(snap.child(currentUID).child(0).val() == actualEventId){
+                                    assistingToEvent = true;
+                                }
+                            }
+                        });
+                        //Inscripción al evento del usuario
+                        if(!assistingToEvent){
+                            var eventXusersKey = firebase.database().ref("eventsXusers/").push().key;
+                            var refEventsXusers = firebase.database().ref("eventsXusers/"+eventXusersKey);
+                            refEventsXusers.set({
+                                [currentUID]: [actualEventId]
+                            });
+
+                            //Se reducen los cupos disponibles del evento
+                            firebase.database().ref("/events").child(actualEventId).child("numberOfRemainingAsisstants").set(event.val().numberOfRemainingAsisstants - 1);
+
+                            alert("Se ha inscrito en el evento correctamente");
+                        } else {
+                            alert("Ya se encuentra inscrito para este evento correctamente");
+                        }
+                    });
+                }
+            })
+        } else {
+            alert("En este momento no te encuentras autenticado en el sistema, por favor realiza la autenticación");
+        }
+    }
+
     render () {  
         return (
             <section>
@@ -40,7 +90,7 @@ class ViewEvent extends React.Component {
                     <h2 className="home">{this.state.actualEvent.name}</h2>
                 </article>
                 <article className="App-row-elements">
-                    <img className="image-left image-event" src={this.state.actualEvent.ulrImage} alt="Event Image"/>
+                    <img className="image-left image-event" src={this.state.actualEvent.urlImage} alt="Event Image"/>
                     <article className="App-column-elements App-center-elements App-font-event">
                         <div className="center-text">
                             <h3>Fecha</h3>{this.state.actualEvent.date}
@@ -57,8 +107,10 @@ class ViewEvent extends React.Component {
                         <div className="center-text">
                             <h3>Número de asistentes</h3>{`${this.state.actualEvent.numberOfRemainingAsisstants}/${this.state.actualEvent.numberOfTotalAssistants}`}
                         </div>
+                        <div className="center-text">
+                            <Button className={"App-button App-button-view"} onClick={this.assistToAnEvent} buttonInfo="Asistir a Evento"></Button>
+                        </div>
                     </article>
-                    
                 </article>
             </section>
         );
